Refactor unlock condition handling by creating a dedicated formatCondition utility, improving readability and reusability across LockedEmployeeCard and LockedOpenSourceCard components.
