+++ client/src/components/LockedEmployeeCard.jsx
@@ -11,22 +11,9 @@
 
 import { formatMoney } from "../utils/currency";
 import { ProgressBar } from "./ProgressBar.jsx";
+import { formatCondition } from "../utils/formatCondition.js";
 
 export function LockedEmployeeCard({ config, progress }) {
-  // Format unlock condition for human-readable display
-  const formatCondition = (condition) => {
-    switch (condition.type) {
-      case "TOTAL_LOC":
-        return `Earn ${condition.required} total lines of code`;
-      case "TOTAL_EMPLOYEE_COUNT":
-        return `Hire ${condition.required} employees`;
-      case "SPECIFIC_EMPLOYEE_COUNT":
-        return `Hire ${condition.required} ${condition.employeeType}${condition.required > 1 ? "s" : ""}`;
-      default:
-        return "Complete an unknown objective";
-    }
-  };
-
   // Calculate if all conditions are met (derived state, not stored)
   // Guard against empty progress array as defensive measure, though in practice
   // this component is only rendered for locked employees with non-empty progress.
