+++ client/src/components/LockedOpenSourceCard.jsx
@@ -1,9 +1,9 @@
 import { formatMoney } from "../utils/currency";
 import { ProgressBar } from "./ProgressBar.jsx";
+import { formatCondition } from "../utils/formatCondition.js";
 
 export function LockedOpenSourceCard({
   project,
-  unlockConditions,
   progress,
   nextBonus,
   locCost,
@@ -49,9 +49,7 @@ export function LockedOpenSourceCard({
                     key={index}
                     current={prog.current}
                     required={prog.required}
-                    label={`Hire ${prog.required} ${
-                      unlockConditions[index].employeeType
-                    }${prog.required > 1 ? "s" : ""}`}
+                    label={formatCondition(prog)}
                   />
                 ))}
               </div>
